requirements.txt
mysql-connector-python
aiomysql
faker